Image_extract/AI/requirements.txt
flask == 2.3.2
flask_cors == 3.0.10
opencv-python == 4.7.0.72
scikit-learn == 1.2.2
numpy == 1.24.3
cmake == 3.26.3
dlib == 19.24.1
face_recognition == 1.3.0
pyzbar == 0.1.9
boto3 == 1.26.134
load_dotenv == 0.1.0
requests == 2.30.0
gunicorn==20.1.0

# pip install -r re...txt
